+++ app/Providers/AuthServiceProvider.php
@@ -4,9 +4,11 @@ namespace App\Providers;
 
 use App\Models\Contact;
 use App\Models\KbArticle;
+use App\Models\Message;
 use App\Models\Ticket;
 use App\Policies\ContactPolicy;
 use App\Policies\KbArticlePolicy;
+use App\Policies\MessagePolicy;
 use App\Policies\TicketPolicy;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\Gate;
@@ -17,6 +19,7 @@ class AuthServiceProvider extends ServiceProvider
         Ticket::class => TicketPolicy::class,
         Contact::class => ContactPolicy::class,
         KbArticle::class => KbArticlePolicy::class,
+        Message::class => MessagePolicy::class,
     ];
 
     public function boot(): void